// src/app/shared/helpers/model/paginationDTO.ts
export class PaginationDTO {
    draw: number = 1;
    length: number = 5;
    start: number = 0;
    sortByKey: string;
    sortByType: string;
    searchWith: string;
    filterType: string;
    totalSize: number = 0;
}
export class PaginationDTO1 {
    draw1: number = 1;
    length1: number = 5;
    start1: number = 0;
    sortByKey: string;
    sortByType: string;
    searchWith: string;
    filterType: string;
    totalSize1: number = 0;
}
export class PaginationDTO2 {
    draw2: number = 1;
    length2: number = 5;
    start2: number = 0;
    sortByKey: string;
    sortByType: string;
    searchWith: string;
    filterType: string;
    totalSize2: number = 0;
}
export class PaginationDTO3 {
    draw3: number = 1;
    length3: number = 5;
    start3: number = 0;
    sortByKey: string;
    sortByType: string;
    searchWith: string;
    filterType: string;
    totalSize3: number = 0;
}
export class PaginationDTO4 {
    draw4: number = 1;
    length4: number = 5;
    start4: number = 0;
    sortByKey: string;
    sortByType: string;
    searchWith: string;
    filterType: string;
    totalSize4: number = 0;
}
export class PaginationDTO5 {
    draw5: number = 1;
    length5: number = 5;
    start5: number = 0;
    sortByKey: string;
    sortByType: string;
    searchWith: string;
    filterType: string;
    totalSize5: number = 0;
}